DRY up the middleware and provider
